MdeModulePkg/EbcDxe: add a stack parameter tracker for ARM
This is intended to fix the CALLEX to native vs ARM calling convention issue by
tracking whether each potential call argument is a natural or a 64-bit value.

This is accomplished by monitoring stack pointer operations, including arithmetic
or single stack buffer realloc, through a separate stack tracker, that grows at
most up to 1/32th the size of the stack. The stack tracker is then used at calltime
to determine 64-bit arguments that need to be aligned.

Note that this only addresses the EBC to native calling convention issue.
It does not address the symmetric native to EBC calling convention issue.